Add game config file
+ Update nob to build raylib for wayland

Unfortunately fullscreen continues to be buggy on Linux whether built
for Wayland or X11, just in a different way :/
However, it seems to be working on Windows – at least if I run the exe
through wine – so that's nice. Actually it's really funny that
fullscreen doesn't work properly on Wayland (native) or X11 (emulating
the display server), but Windows (literally emulating all the system
calls)? No problem!