Add build instructions and fix Windows build
Turns out you need to link to gdi32 and winmm whether cross-compiling or
compiling natively. Anyways, it should work on Windows now. Because the
Code::Blocks project isn't statically linking the exe, there's a good
chance you won't be able to copy-paste the exe onto a machine withouth
MinGW installed and still play the game (it might be dynamically linking
to C/C++ runtime libraries), but that's not the end of the world.